# contracts/openapi.yaml
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/auth/wechat": {
      "post": {
        "operationId": "ApiController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "code"
                ],
                "properties": {
                  "code": {
                    "type": "string",
                    "maxLength": 512
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Session"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "410": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [],
        "summary": "微信 code 换业务会话",
        "tags": [
          "Pidan Vocal v1"
        ]
      }
    },
    "/v1/auth/session": {
      "delete": {
        "operationId": "ApiController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Revoked"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "410": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Pidan Vocal v1"
        ]
      }
    },
    "/v1/me": {
      "get": {
        "operationId": "ApiController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Me"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "410": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Pidan Vocal v1"
        ]
      }
    },
    "/v1/books": {
      "get": {
        "operationId": "ApiController_list",
        "parameters": [
          {
            "name": "audience",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "sample",
                "member"
              ],
              "default": "sample"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contentType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "book",
                "blog",
                "movie",
                "tv"
              ]
            },
            "description": "Omit for all types; cursor is bound to this filter."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/BookPage"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "410": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Business error (409 PROGRESS_CONFLICT details contains ProgressResult)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "tags": [
          "Pidan Vocal v1"
        ]
      }
    },
...
  },
  "info": {
    "title": "Pidan Vocal API",
    "description": "",
    "version": "1.3.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "ready": {
        "type": "http",
        "scheme": "bearer",
        "description": "Production readiness token; separate from user sessions"
      }
    },
    "schemas": {
      "Session": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              }
            },
            "required": [
              "id"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "accessToken",
          "expiresAt",
          "user"
        ],
        "additionalProperties": false
      },
...
      "BookSummary": {
        "type": "object",
        "properties": {
          "seasonCount": {
            "type": "integer",
            "minimum": 0
          },
          "contentType": {
            "type": "string",
            "enum": [
              "book",
              "blog",
              "movie",
              "tv"
            ]
          },
          "unitCount": {
            "type": "integer",
            "minimum": 0
          },
          "coverUrl": {
            "type": "string",
            "format": "uri",
            "pattern": "^https://"
          },
          "bookId": {
            "type": "string"
          },
          "buildId": {
            "type": "string"
          },
          "textRevision": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "edition": {
            "type": "string"
          },
          "contentScope": {
            "type": "string",
            "enum": [
              "sample",
              "complete"
            ]
          },
          "visibility": {
            "type": "string",
            "enum": [
              "sample-public",
              "private"
            ]
          },
          "chapterCount": {
            "type": "integer",
            "minimum": 0
          },
          "contentChapterCount": {
            "type": "integer",
            "minimum": 0
          },
          "sentenceCount": {
            "type": "integer",
            "minimum": 0
          },
          "playableCount": {
            "type": "integer",
            "minimum": 0
          },
          "chapterAudioAvailableCount": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "seasonCount",
          "contentType",
          "unitCount",
          "bookId",
          "buildId",
          "textRevision",
          "title",
          "language",
          "edition",
          "contentScope",
          "visibility",
          "chapterCount",
          "contentChapterCount",
          "sentenceCount",
          "playableCount",
          "chapterAudioAvailableCount"
        ],
        "additionalProperties": false
      },
      "BookPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookSummary"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "additionalProperties": false
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ],
            "additionalProperties": false
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "requestId"
        ],
        "additionalProperties": false
      },
      "Me": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "additionalProperties": false
      },
      "Revoked": {
        "type": "object",
        "properties": {
          "revoked": {
            "type": "boolean"
          }
        },
        "required": [
          "revoked"
        ],
        "additionalProperties": false
      },
...
    }
  }
}